+++ FE/src/components/CompanyDetailPage/DeleteAndChangeModal/DeleteAndChangeModal.module.scss
@@ -1,0 +1,105 @@
+.DeleteAndChangeModal {
+  width: 9rem; /* 144px */
+  height: 5rem; /* 80px */
+  display: flex;
+  flex-direction: column;
+  justify-content: space-between;
+}
+
+.DeleteAndChangeModalButton1 {
+  width: 100%;
+  height: 2.5rem; /* 40px */
+  border-top-left-radius: 0.625rem; /* 10px */
+  border-top-right-radius: 0.625rem; /* 10px */
+  border: 1px solid #747474;
+  background-color: #131313;
+  font-family: Pretendard, sans-serif;
+  font-weight: 400;
+  font-size: 0.875rem; /* 14px */
+  color: #d8d8d8;
+}
+
+.DeleteAndChangeModalButton2 {
+  width: 100%;
+  height: 2.5rem; /* 40px */
+  border-bottom-left-radius: 0.625rem; /* 10px */
+  border-bottom-right-radius: 0.625rem; /* 10px */
+  border: 1px solid #747474;
+  background-color: #131313;
+  font-family: Pretendard, sans-serif;
+  font-weight: 400;
+  font-size: 0.875rem; /* 14px */
+  color: #d8d8d8;
+}
+
+/* PC (고정 120rem, 1920px) */
+@media screen and (min-width: 120rem) {
+  .companyDetailDescription {
+    max-width: 120rem;
+    margin: 0 auto;
+  }
+  .CompanyDetailPage,
+  .CompanyDetailDiv,
+  .ViewMyStartUpDiv,
+  .companyDetailThreePart {
+    max-width: 120rem;
+  }
+}
+
+/* Tablet (46.5rem ~ 74.9375rem, 744px ~ 1199px) */
+@media screen and (min-width: 46.5rem) and (max-width: 74.9375rem) {
+  .companyDetailDescription {
+    height: auto;
+    padding: 1.25rem;
+  }
+  .detialPageWeight400, .detialPageWeight600 {
+    font-size: 0.875rem;
+    line-height: 1.125rem;
+  }
+  .CompanyDetailDiv,
+  .ViewMyStartUpDiv {
+    max-width: 100%;
+  }
+  .DeleteAndChangeModal {
+    width: 80%;
+    height: auto;
+    padding: 1rem;
+  }
+  .DeleteAndChangeModalButton1, .DeleteAndChangeModalButton2 {
+    font-size: 1rem;
+  }
+}
+
+/* Tablet (744px ~ 1199px) */
+@media screen and (min-width: 46.5rem) and (max-width: 74.9375rem) {
+  .DeleteAndChangeModal {
+    width: 80%;
+    height: auto;
+    padding: 1rem;
+  }
+  .DeleteAndChangeModalButton1, .DeleteAndChangeModalButton2 {
+    font-size: 1rem;
+  }
+}
+
+/* Mobile (375px ~ 743px) */
+@media screen and (min-width: 23.4375rem) and (max-width: 46.4375rem) {
+  .DeleteAndChangeModal {
+    width: 90%;
+    height: auto;
+  }
+  .DeleteAndChangeModalButton1, .DeleteAndChangeModalButton2 {
+    font-size: 1rem;
+  }
+}
+
+/* Very Small Mobile (최대 374px) */
+@media screen and (max-width: 23.375rem) {
+  .DeleteAndChangeModal {
+    width: 95%;
+    height: auto;
+  }
+  .DeleteAndChangeModalButton1, .DeleteAndChangeModalButton2 {
+    font-size: 1rem;
+  }
+}
