+++ FE/src/components/CompanyDetailPage/Description/Description.jsx
@@ -1,9 +1,9 @@
-import styles from "../CompanyDetailPage.module.scss";
+import styles from "./Description.module.scss";
 
-export function Description({ className, text }) {
+export function Description({ text }) {
   return (
     <>
-      <div className={className}>
+      <div className={styles.companyDetailDescription}>
         <p className={styles.detialPageWeight600}>기업 소개</p>
         <p className={styles.detialPageWeight400}>{text}</p>
       </div>
