+++ FE/src/components/CompanyDetailPage/Description/Description.module.scss
@@ -1,0 +1,73 @@
+.companyDetailDescription {
+  max-width: 75rem;
+  width: 100%;
+  height: 287px;
+  border-radius: 10px;
+  gap: 16px;
+  padding: 24px;
+  background-color: #212121;
+  display: flex;
+  flex-direction: column;
+}
+
+.detialPageWeight400 {
+  font-family: Pretendard, sans-serif;
+  font-weight: 400;
+  font-size: 16px;
+  line-height: 20px;
+  letter-spacing: 0;
+  color: #d8d8d8;
+}
+
+.detialPageWeight600 {
+  font-family: Pretendard, sans-serif;
+  font-weight: 600;
+  font-size: 16px;
+  line-height: 100%;
+  letter-spacing: 0;
+}
+
+/* PC (고정 120rem, 1920px) */
+@media screen and (min-width: 120rem) {
+  .companyDetailDescription {
+    max-width: 120rem;
+    margin: 0 auto;
+  }
+}
+
+/* Tablet (46.5rem ~ 74.9375rem, 744px ~ 1199px) */
+@media screen and (min-width: 46.5rem) and (max-width: 74.9375rem) {
+  .companyDetailDescription {
+    height: auto;
+    padding: 1.25rem;
+  }
+  .detialPageWeight400, .detialPageWeight600 {
+    font-size: 0.875rem;
+    line-height: 1.125rem;
+  }
+}
+
+/* Mobile (23.4375rem ~ 46.4375rem, 375px ~ 743px) */
+@media screen and (min-width: 23.4375rem) and (max-width: 46.4375rem) {
+  .companyDetailDescription {
+    padding: 1rem;
+    overflow-x: auto;
+    white-space: nowrap;
+  }
+  .detialPageWeight400, .detialPageWeight600 {
+    font-size: 0.8125rem;
+    line-height: 1.0625rem;
+  }
+}
+
+/* Very Small Mobile (최대 23.375rem, 최대 374px) */
+@media screen and (max-width: 23.375rem) {
+  .companyDetailDescription {
+    padding: 0.75rem;
+    overflow-x: auto;
+  }
+  .detialPageWeight400, .detialPageWeight600 {
+    font-size: 0.75rem;
+    line-height: 1rem;
+  }
+}
