+++ FE/src/components/CompanyDetailPage/InvestChangeModal/InvestChangeModal.jsx
@@ -1,4 +1,4 @@
-import styles from "../CompanyDetailPage.module.scss";
+import styles from "./InvestChangeModal.module.scss";
 
 export function InvestChangeModal({ modalChangeState }) {
   return (
