+++ FE/src/components/CompanyDetailPage/InvestTable/InvestTable.module.scss
@@ -1,0 +1,159 @@
+.investTableHeader {
+  max-width: 75rem; /* 1200px */
+  width: 100%;
+  height: 2.5rem; /* 40px */
+  gap: 0.625rem; /* 10px */
+  background-color: #2e2e2e;
+  padding: 0 1rem; /* 16px */
+  border-spacing: 0;
+  border-radius: 0.25rem; /* 4px */
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  flex-direction: row;
+  margin: 0 auto;
+}
+
+.detialPageWeight500 {
+  font-family: Pretendard, sans-serif;
+  font-weight: 500;
+  font-size: 1.25rem; /* 20px */
+  line-height: 100%;
+  letter-spacing: 0;
+}
+
+.investTableBody {
+  width: 100%;
+  gap: 0.625rem; /* 10px */
+  background-color: #2e2e2e;
+  padding: 0 1rem; /* 16px */
+  border-spacing: 0;
+  border-radius: 0.25rem;
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  flex-direction: column;
+}
+
+.investTableRow {
+  width: 100%;
+  height: 4rem; /* 64px */
+  gap: 0.625rem;
+  border-bottom-width: 1px;
+  padding: 0 1rem;
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  background-color: #2e2e2e;
+  border-bottom: 1px solid #4b4b4b;
+}
+
+.detialPageWeight400 {
+  font-family: Pretendard, sans-serif;
+  font-weight: 400;
+  font-size: 1rem; /* 16px */
+  line-height: 1.25rem; /* 20px */
+  letter-spacing: 0;
+  color: #d8d8d8;
+}
+
+.modalOverlay {
+  position: fixed;
+  top: 0;
+  left: 0;
+  width: 100vw;
+  height: 100vh;
+  background: rgba(0, 0, 0, 0.5);
+  display: flex;
+  justify-content: center;
+  align-items: center;
+}
+
+.investDeleteButton {
+  width: 1.5rem; /* 24px */
+  height: 1.5rem;
+  background-image: url("../../../pages/CompanyDetailPage/photo/ic_kebab.png");
+}
+
+/* 반응형 스타일 */
+
+/* PC (1920px 고정) */
+@media screen and (min-width: 120rem) {
+  .investTableHeader {
+    max-width: 75rem; /* 1200px */
+    height: 2.5rem; /* 40px */
+    padding: 0 1rem; /* 16px */
+  }
+
+  .investTableRow {
+    height: 4rem; /* 64px */
+    padding: 0 1rem;
+  }
+
+  .detialPageWeight500 {
+    font-size: 1.25rem; /* 20px */
+  }
+
+  .detialPageWeight400 {
+    font-size: 1rem; /* 16px */
+  }
+}
+
+/* Tablet (744px ~ 1199px) */
+@media screen and (min-width: 46.5rem) and (max-width: 74.9375rem) {
+  .investTableHeader {
+    max-width: 100%;
+  }
+
+  .investTableRow {
+    flex-direction: row;
+    height: auto;
+    padding: 1rem;
+  }
+}
+
+/* Mobile (375px ~ 743px) */
+@media screen and (min-width: 23.4375rem) and (max-width: 46.4375rem) {
+  .investTableHeader {
+    width: 100%;
+    flex-direction: row;
+    gap: 0.5rem;
+    text-align: center;
+  }
+
+  .investTableRow {
+    width: 100%;
+    padding: 0.5rem;
+  }
+
+  .detialPageWeight500 {
+    font-size: 1rem; /* 16px */
+  }
+
+  .detialPageWeight400 {
+    font-size: 0.875rem; /* 14px */
+  }
+}
+
+/* Very Small Mobile (최대 374px) */
+@media screen and (max-width: 23.375rem) {
+  .investTableHeader {
+    width: 100%;
+    flex-direction: row;
+    gap: 0.5rem;
+    text-align: center;
+  }
+
+  .investTableRow {
+    width: 100%;
+    padding: 0.5rem;
+  }
+
+  .detialPageWeight500 {
+    font-size: 1rem; /* 16px */
+  }
+
+  .detialPageWeight400 {
+    font-size: 0.875rem; /* 14px */
+  }
+}
