+++ FE/src/components/CompanyDetailPage/PaseNationButton/PaseNationButton.module.scss
@@ -1,0 +1,56 @@
+.PaseNationButton {
+  width: 3rem; /* 48px */
+  height: 3rem;
+  border-radius: 0.625rem; /* 10px */
+  font-family: Pretendard, sans-serif;
+  font-weight: 400;
+  font-size: 1.125rem; /* 18px */
+  line-height: 100%;
+  letter-spacing: 0;
+  text-align: center;
+  background-color: #2e2e2e;
+}
+
+/* PC 1920px 고정 */
+@media (min-width: 120rem) {
+  .PaseNationButton {
+    width: 3rem; /* 48px */
+    height: 3rem;
+    font-size: 1.125rem; /* 18px */
+  }
+}
+
+/* 태블릿 화면 (744px ~ 1199px) */
+@media (min-width: 46.5rem) and (max-width: 74.9375rem) {
+  .PaseNationButton {
+    width: 2.75rem; /* 44px */
+    height: 2.75rem; /* 44px */
+    font-size: 1.05rem; /* 16.8px */
+  }
+}
+
+/* 모바일 화면 (375px ~ 743px) */
+@media (min-width: 23.4375rem) and (max-width: 46.5rem) {
+  body {
+    overflow-x: auto; /* 가로 스크롤을 강제로 발생 */
+  }
+
+  .PaseNationButton {
+    width: 2.5rem; /* 40px */
+    height: 2.5rem; /* 40px */
+    font-size: 1rem; /* 16px */
+  }
+}
+
+/* 가장 작은 화면 (375px 이하) */
+@media (max-width: 23.4375rem) {
+  body {
+    overflow-x: auto; /* 가로 스크롤을 강제로 발생 */
+  }
+
+  .PaseNationButton {
+    width: 2.25rem; /* 36px */
+    height: 2.25rem; /* 36px */
+    font-size: 0.875rem; /* 14px */
+  }
+}
